Presentation on Wavecon Telecom Analysis
my presentation on Wavecon Telecom Analysis completed as part of the Codebasics Virtual Internship!

In this project, I analyzed the impact of the 5G launch on key business metrics and plans. Using Power BI and telecom KPIs, I identified:

 📉 A slight revenue decline post-5G launch

 📊 Underperforming and discontinued plans

 📈 Strong performance of entry-level recharge plans (p1 & p2)

 🧠 Strategic recommendations for plan restructuring and 5G-focused growth

This hands-on project sharpened my skills in business analysis, storytelling with data, and dashboard interpretation.